Refactor and Organize
