+++ Hifdh Tracker/Lauch/AppDelegate.swift
@@ -7,18 +7,15 @@
 
 import UIKit
 import CoreData
-enum UserDefaultsKey: String {
-    case isFirstRun = "isFirstRun"
-    case isFromFront = "isFromFront"
-}
+
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
     
     let userDefaults = UserDefaults.standard
-
+    
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
-        
+        // MARK: Init UserDefaults booleans
         // set default first run values
         let defaultValues = [UserDefaultsKey.isFirstRun.rawValue : true, UserDefaultsKey.isFromFront.rawValue : true]
         userDefaults.register(defaults: defaultValues)
@@ -27,41 +24,41 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // if first run
         if userDefaults.bool(forKey: UserDefaultsKey.isFirstRun.rawValue){
             Page.getDefaultPages(managedContext)
-           
+            
             userDefaults.set(false, forKey: UserDefaultsKey.isFirstRun.rawValue)
         }
         return true
     }
-
+    
     // MARK: UISceneSession Lifecycle
-
+    
     func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
         // Called when a new scene session is being created.
         // Use this method to select a configuration to create the new scene with.
         return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
     }
-
+    
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
         // Called when the user discards a scene session.
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
-
+    
     // MARK: - Core Data stack
-
+    
     lazy var persistentContainer: NSPersistentContainer = {
         /*
          The persistent container for the application. This implementation
          creates and returns a container, having loaded the store for the
          application to it. This property is optional since there are legitimate
          error conditions that could cause the creation of the store to fail.
-        */
+         */
         let container = NSPersistentContainer(name: "Hifdh_Tracker")
         container.loadPersistentStores(completionHandler: { (storeDescription, error) in
             if let error = error as NSError? {
                 // Replace this implementation with code to handle the error appropriately.
                 // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
-                 
+                
                 /*
                  Typical reasons for an error here include:
                  * The parent directory does not exist, cannot be created, or disallows writing.
@@ -77,7 +74,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     }()
     
     // MARK: - Core Data Saving support
-
+    
     func saveContext () {
         let context = persistentContainer.viewContext
         if context.hasChanges {
@@ -91,6 +88,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             }
         }
     }
+}
 
+// MARK: Enumerations
+enum UserDefaultsKey: String {
+    case isFirstRun = "isFirstRun"
+    case isFromFront = "isFromFront"
 }
