+++ Hifdh Tracker/Log/AddLogViewController.swift
@@ -10,17 +10,18 @@ import CoreData
 
 class AddLogViewController: UIViewController {
     
+    // MARK: Outlets
     @IBOutlet weak var fromPageTextField: UITextField!
     @IBOutlet weak var untilPageTextField: UITextField!
     @IBOutlet weak var datePicker: UIDatePicker!
-    
     @IBOutlet weak var errorLabel: UILabel!
     @IBOutlet weak var hideErrorConstraint: NSLayoutConstraint!
     @IBOutlet weak var showErrorConstraint: NSLayoutConstraint!
-    // give access to appdelegate and logs
-    var delegate: AppDelegate?
     
-    //reload on dismiss
+    // MARK: Variables
+    var delegate: AppDelegate? // give access to appdelegate through segue
+    
+    // MARK: VC's onDismiss
     var isDismissed: ( (Int) -> Void )?
     
     func dismiss(andScrollTo pageNumber: Int) {
@@ -28,13 +29,7 @@ class AddLogViewController: UIViewController {
         self.dismiss(animated: true)
     }
     
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        // Do any additional setup after loading the view.
-        
-    }
-    
+    // MARK: Actions
     @IBAction func createLogButtonTouchUpInside(_ sender: Any) {
         
         //take values from "from" and "to" text fields and make them ints
