+++ Hifdh Tracker/Models/Page+CoreDataProperties.swift
@@ -9,27 +9,22 @@
 import CoreData
 import UIKit
 
-
-enum Statistic: Int {
-    case completionDate = 0
-    case pagesPerDay = 1
-    case pagesMemorized = 2
-    case percentMemorized = 3
-}
-
+// MARK: Page Model
 extension Page {
-
+    
     @nonobjc public class func fetchRequest() -> NSFetchRequest<Page> {
         return NSFetchRequest<Page>(entityName: "Page")
     }
-
+    
     @NSManaged public var dateMemorized: Date?
     @NSManaged public var isMemorized: Bool
     @NSManaged public var pageNumber: Int16
-
+    
 }
 
 extension Page : Identifiable {
+    
+    // MARK: custom init
     convenience init(_ pageNumber: Int16, _ context: NSManagedObjectContext) {
         self.init(context: context)
         self.pageNumber = pageNumber
@@ -37,6 +32,7 @@ extension Page : Identifiable {
         self.dateMemorized = nil
     }
     
+    // MARK: Global array
     static var logs: [Page] {
         get {
             let delegate = UIApplication.shared.delegate as! AppDelegate
@@ -44,6 +40,24 @@ extension Page : Identifiable {
         }
     }
     
+    // MARK: Delete all data
+    static func deleteAll(in context: NSManagedObjectContext) {
+        // Initialize Fetch Request
+        let request: NSFetchRequest<Page> = Page.fetchRequest()
+        
+        do {
+            let items = try context.fetch(request)
+            for item in items {
+                context.delete(item)
+            }
+            try context.save()
+            
+        } catch {
+            // TODO: Error Handling
+        }
+    }
+    
+    // MARK: Reset all data
     static func getDefaultPages(_ context: NSManagedObjectContext) -> Void {
         Page.deleteAll(in: context)
         try? context.save()
@@ -53,6 +67,7 @@ extension Page : Identifiable {
         }
     }
     
+    // MARK: Fetch all in context
     static func fetchPages(in context: NSManagedObjectContext) -> [Page] {
         let request: NSFetchRequest<Page> = Page.fetchRequest()
         request.sortDescriptors = [NSSortDescriptor(key: "pageNumber", ascending: (UserDefaults.standard.bool(forKey: UserDefaultsKey.isFromFront.rawValue)))]
@@ -61,6 +76,12 @@ extension Page : Identifiable {
             return pageLogsFromCoreData
         }
         return []
+    }
+    
+    // MARK: Statistics and calculations
+    
+    static var firstNotInMemoryIndexPath: IndexPath {
+        return IndexPath(row: Page.logs.firstIndex(where: { !$0.isMemorized }) ?? 0, section: 1)
     }
     
     static var selectedStat: Statistic?
@@ -100,52 +121,20 @@ extension Page : Identifiable {
     static var completionDate: Date {
         return(Date().advanced(by: ((numberOfNotMemorized/pagesPerDay).convert(from: .days, to: .seconds))))
     }
-    
-    static func deleteAll(in context: NSManagedObjectContext) {
-        // Initialize Fetch Request
-        let request: NSFetchRequest<Page> = Page.fetchRequest()
-        
-        do {
-            let items = try context.fetch(request)
-            for item in items {
-                context.delete(item)
-            }
-            try context.save()
-            
-        } catch {
-            // TODO: Error Handling
-        }
-    }
 }
+
+// MARK: Enumerations
+
+enum Statistic: Int {
+    case completionDate = 0
+    case pagesPerDay = 1
+    case pagesMemorized = 2
+    case percentMemorized = 3
+}
+
 enum TimeUnits {
     case seconds
     case days
     case months
     case years
-}
-extension TimeInterval {
-    func convert(from fromUnit: TimeUnits = .seconds, to toUnit: TimeUnits) -> TimeInterval {
-        var x = self
-        switch fromUnit {
-            case .seconds:
-                break
-            case .days:
-                x *= (60*60*24)
-            case .months:
-                x *= (60*60*24*30)
-            case .years:
-                x *= (60*60*24*365)
-        }
-        switch toUnit {
-            case .seconds:
-                break
-            case .days:
-                x /= (60*60*24)
-            case .months:
-                x /= (60*60*24*30)
-            case .years:
-                x /= (60*60*24*365)
-        }
-        return x
-    }
 }
