+++ Hifdh Tracker/Profile/ProfileViewController.swift
@@ -10,23 +10,15 @@ import CoreData
 
 class ProfileViewController: UIViewController, UIPickerViewDataSource, UIPickerViewDelegate {
     
-    let mainStatistics = ["Predicted Hifdh Completion", "Pages per Day", "Pages Memorized", "Percent Memorized",]
-    
-    func numberOfComponents(in pickerView: UIPickerView) -> Int {
-        return 1
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return mainStatistics.count
-    }
-    
-    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
-        return mainStatistics[row]
-    }
-    
-    
-    var isDropDownExpanded: Bool = false
+    // MARK: Constants
     let delegate = UIApplication.shared.delegate as? AppDelegate
+    
+    
+    // MARK: Variables
+    var isDropDownExpanded: Bool = false
+    
+    
+    // MARK: Outlets
     @IBOutlet weak var mainStatPicker: UIPickerView!
     @IBOutlet weak var mainStatTitleLabel: UILabel!
     @IBOutlet weak var mainStatValueLabel: UILabel!
@@ -34,23 +26,32 @@ class ProfileViewController: UIViewController, UIPickerViewDataSource, UIPickerV
     @IBOutlet weak var memorizationProgressbar: UIProgressView!
     @IBOutlet weak var mainStatPickerHeight: NSLayoutConstraint!
     
+    // MARK: ViewDid...
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         // Do any additional setup after loading the view.
         view.layoutIfNeeded()
         configureProgressBar()
-        configureMainStatCard(for: Page.selectedStat ?? .pagesMemorized)
         configureStatPicker()
     }
     
-    func configureStatPicker() {
+    override func viewWillAppear(_ animated: Bool) {
+        // update all views
+        configureMainStatCard(for: Page.selectedStat ?? .pagesMemorized)
+    }
+    
+    // MARK: Configurations
+    
+    private func configureStatPicker() {
         mainStatPicker.dataSource = self
         mainStatPicker.delegate = self
+        mainStatPicker.selectRow(Statistic.pagesMemorized.rawValue, inComponent: 0, animated: false)
         setMainStatPickerVisibility(to: isDropDownExpanded)
     }
     
-    func configureProgressBar() {
+    private func configureProgressBar() {
         var percentMemorized: Float = 0.0
         withCoreData {
             percentMemorized = Float(Page.numberOfMemorized) / 604.0
@@ -58,7 +59,7 @@ class ProfileViewController: UIViewController, UIPickerViewDataSource, UIPickerV
         memorizationProgressbar.progress = percentMemorized
     }
     
-    func configureMainStatCard(for statistic: Statistic) -> Void {
+    private func configureMainStatCard(for statistic: Statistic) -> Void {
         
         switch statistic {
             case .completionDate:
@@ -90,14 +91,13 @@ class ProfileViewController: UIViewController, UIPickerViewDataSource, UIPickerV
         }
     }
     
-    
-    
+    // MARK: Actions
     @IBAction func statisticsDropDownButtonPressed(_ sender: UIButton) {
         var imageName = "chevron."
         if isDropDownExpanded {
             isDropDownExpanded = false
             imageName += "down"
-            configureMainStatCard(for: Statistic(rawValue:(mainStatPicker.selectedRow(inComponent: 0))) ?? .pagesMemorized)
+            Page.selectedStat = Statistic(rawValue: mainStatPicker.selectedRow(inComponent: 0))
             
         } else {
             isDropDownExpanded = true
@@ -107,28 +107,40 @@ class ProfileViewController: UIViewController, UIPickerViewDataSource, UIPickerV
         setMainStatPickerVisibility(to: isDropDownExpanded)
     }
     
+    // MARK: Animations / Dynamic Constraints
     private func setMainStatPickerVisibility(to show: Bool) {
         
         UIView.animate(withDuration: 0.3){ [self] in
-        if show {
-            mainStatPickerHeight.constant = 100
-            self.view.layoutIfNeeded()
-        } else {
-            mainStatPickerHeight.constant = 0
-            self.view.layoutIfNeeded()
-            
-        }
+            if show {
+                mainStatPickerHeight.constant = 100
+                self.view.layoutIfNeeded()
+            } else {
+                mainStatPickerHeight.constant = 0
+                self.view.layoutIfNeeded()
+                
+            }
         }
     }
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
+    
+    // MARK: UIPicker
+    
+    let mainStatistics = ["Predicted Hifdh Completion", "Pages per Day", "Pages Memorized", "Percent Memorized",]
+    
+    func numberOfComponents(in pickerView: UIPickerView) -> Int {
+        return 1
     }
-    */
+    
+    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
+        return mainStatistics.count
+    }
+    
+    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int, forComponent component: Int) -> String? {
+        return mainStatistics[row]
+    }
+    func pickerView(_ pickerView: UIPickerView, didSelectRow row: Int, inComponent component: Int) {
+        Page.selectedStat = Statistic(rawValue: mainStatPicker.selectedRow(inComponent: 0))
+        configureMainStatCard(for: Page.selectedStat ?? .pagesMemorized)
+    }
     
     // MARK: Core Data
     
@@ -138,5 +150,14 @@ class ProfileViewController: UIViewController, UIPickerViewDataSource, UIPickerV
         }
         delegate?.saveContext()
     }
-
+    
+    /*
+     // MARK: - Navigation
+     
+     // In a storyboard-based application, you will often want to do a little preparation before navigation
+     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+     // Get the new view controller using segue.destination.
+     // Pass the selected object to the new view controller.
+     }
+     */
 }
